typescript

+++ src/util/logger.ts
@@ -16,8 +16,7 @@ export function pushLog(level: LogLevel, ...args: unknown[]): void {
       try {
         return JSON.stringify(v);
       } catch {
-        // eslint-disable-next-line @typescript-eslint/no-base-to-string
-        return v?.toString();
+        return (v as bigint)?.toString();
       }
     }).join(", "),
   });
